--- components/Rebalance/RebalanceController.tsx
@@ -1,8 +1,0 @@
-import * as React from 'react'
-import RebalanceView from './RebalanceView'
-
-const RebalanceController = () => {
-  return <RebalanceView />
-}
-
-export default RebalanceController
